Moving trivial class into using class, and update tests for comments
