+++ collector-kinesis/pom.xml
@@ -48,7 +48,7 @@
       <artifactId>awaitility</artifactId>
       <scope>test</scope>
     </dependency>
-
+    
     <dependency>
       <groupId>${project.groupId}</groupId>
       <artifactId>zipkin-aws-junit</artifactId>
