--- collector-kinesis/src/main/java/zipkin/collector/kinesis/KinesisRecordProcessorFactory.java
@@ -1,32 +1,0 @@
-/**
- * Copyright 2016-2017 The OpenZipkin Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-package zipkin.collector.kinesis;
-
-import com.amazonaws.services.kinesis.clientlibrary.interfaces.v2.IRecordProcessor;
-import com.amazonaws.services.kinesis.clientlibrary.interfaces.v2.IRecordProcessorFactory;
-import zipkin.collector.Collector;
-
-public class KinesisRecordProcessorFactory implements IRecordProcessorFactory {
-
-  private final Collector collector;
-
-  KinesisRecordProcessorFactory(Collector collector){
-    this.collector = collector;
-  }
-
-  @Override
-  public IRecordProcessor createProcessor() {
-    return new KinesisSpanProcessor(collector);
-  }
-}
